refactor: fix problems related to imports
This commit is runnable.
